[CPDLP-1564] - refactor under app/services
query + presenter objects
