+++ app/controllers/finance/ecf/assurance_reports_controller.rb
@@ -1,66 +1,14 @@
 # frozen_string_literal: true
 
+require "csv_serialiser"
+
 module Finance
   module ECF
-    class AssuranceReportsController < BaseController
-      def show
-        send_data(generate_csv, filename:)
-      end
-
+    class AssuranceReportsController < Finance::AssuranceReportsController
     private
 
-      def generate_csv
-        CSV.generate do |csv|
-          csv << csv_headers
-          assurance_report_rows.each do |assurance_report_row|
-            csv << assurance_report_row.to_csv
-          end
-        end
-      end
-
-      def assurance_report_rows
-        @assurance_report_rows ||= AssuranceReportQuery.new(statement).rows
-      end
-
-      def csv_headers
-        [
-          "Participant ID",
-          "Participant Name",
-          "TRN",
-          "Type",
-          "Mentor Profile ID",
-          "Schedule",
-          "Eligible For Funding",
-          "Eligible For Funding Reason",
-          "Sparsity Uplift",
-          "Pupil Premium Uplift",
-          "Sparsity And Pp",
-          "Lead Provider Name",
-          "Delivery Partner Name",
-          "School Urn",
-          "School Name",
-          "Training Status",
-          "Training Status Reason",
-          "Declaration ID",
-          "Declaration Status",
-          "Declaration Type",
-          "Declaration Date",
-          "Declaration Created At",
-          "Statement Name",
-          "Statement ID",
-        ]
-      end
-
-      def filename
-        "ECF-Declarations-#{lead_provider.name.gsub(/\W/, '')}-Cohort#{statement.cohort.start_year}-#{statement.name.gsub(/\W/, '')}.csv"
-      end
-
-      def lead_provider
-        @statement.lead_provider
-      end
-
-      def statement
-        @statement ||= Finance::Statement::ECF.find(params[:statement_id])
+      def assurance_report_presenter
+        @assurance_report_presenter ||= AssuranceReport::Presenter.new(params[:statement_id])
       end
     end
   end
