+++ app/controllers/finance/npq/assurance_reports_controller.rb
@@ -1,60 +1,14 @@
 # frozen_string_literal: true
 
+require "csv_serialiser"
+
 module Finance
   module NPQ
-    class AssuranceReportsController < BaseController
-      def show
-        send_data(generate_csv, filename:)
-      end
-
+    class AssuranceReportsController < Finance::AssuranceReportsController
     private
 
-      def generate_csv
-        CSV.generate do |csv|
-          csv << csv_headers
-          assurance_report_rows.each do |assurance_report_row|
-            csv << assurance_report_row.to_csv
-          end
-        end
-      end
-
-      def assurance_report_rows
-        @assurance_report_rows ||= AssuranceReportQuery.new(statement).rows
-      end
-
-      def csv_headers
-        [
-          "Participant ID",
-          "Participant Name",
-          "TRN",
-          "Course Identifier",
-          "Schedule",
-          "Eligible For Funding",
-          "Lead Provider Name",
-          "School Urn",
-          "School Name",
-          "Training Status",
-          "Training Status Reason",
-          "Declaration ID",
-          "Declaration Status",
-          "Declaration Type",
-          "Declaration Date",
-          "Declaration Created At",
-          "Statement Name",
-          "Statement ID",
-        ]
-      end
-
-      def filename
-        "NPQ-Declarations-#{npq_lead_provider.name.gsub(/\W/, '')}-Cohort#{statement.cohort.start_year}-#{statement.name.gsub(/\W/, '')}.csv"
-      end
-
-      def npq_lead_provider
-        statement.npq_lead_provider
-      end
-
-      def statement
-        @statement ||= Finance::Statement::NPQ.find(params[:statement_id])
+      def assurance_report_presenter
+        @assurance_report_presenter ||= AssuranceReport::Presenter.new(params[:statement_id])
       end
     end
   end
