+++ app/services/finance/npq/assurance_report.rb
@@ -6,6 +6,8 @@ module Finance
       include ActiveModel::Model
       include ActiveModel::Attributes
 
+      attribute :id
+      attribute :npq_lead_provider_id
       attribute :participant_id
       attribute :participant_name
       attribute :trn
